refactoring for PR
